+++ src/components/CustomInputs/JsonInput.vue
@@ -184,7 +184,7 @@ export default {
       <v-icon :color="iconColor">
         {{ prependIcon }}
       </v-icon>
-      <div class="caption o-20">
+      <div class="text-caption o-20">
         {{ prependIconLabel }}
       </div>
     </div>
@@ -224,7 +224,7 @@ export default {
       Format
     </v-btn>
 
-    <div class="caption red--text min-height pl-4">{{ jsonError }}</div>
+    <div class="text-caption red--text min-height pl-4">{{ jsonError }}</div>
   </div>
 </template>