+++ src/pages/FlowRun/TaskRunTable-Tile.vue
@@ -47,7 +47,13 @@ export default {
           width: '20%'
         },
         { text: 'End Time', value: 'end_time', align: 'start', width: '15%' },
-        { text: 'Duration', value: 'duration', align: 'end', width: '17.5%' },
+        {
+          text: 'Duration',
+          value: 'duration',
+          align: 'end',
+          width: '17.5%',
+          sortable: false
+        },
         { text: 'State', value: 'state', align: 'end', width: '10%' },
         {
           text: '',
@@ -177,7 +183,7 @@ export default {
             >
               {{ item }}
             </v-chip>
-            <span v-if="index === 2" class="grey--text caption">
+            <span v-if="index === 2" class="grey--text text-caption">
               (+{{ state.length - 2 }})
             </span>
           </template>
@@ -219,7 +225,7 @@ export default {
         :server-items-length="taskRunsCount"
         :sort-by.sync="sortBy"
         :sort-desc.sync="sortDesc"
-        :class="{ 'fixed-table': this.$vuetify.breakpoint.smAndUp }"
+        :class="{ 'fixed-table': $vuetify.breakpoint.smAndUp }"
         calculate-widths
       >
         <template #item.task.name="{ item }">