+++ src/pages/Notifications/NotificationGroup.vue
@@ -224,7 +224,7 @@ export default {
 
 <template>
   <v-card v-show="notificationsCount > 0" class="mb-4 py-2 px-4" tile>
-    <v-card-title class="pl-1 subtitle-2 grey--text text--darken-2">
+    <v-card-title class="pl-1 text-subtitle-2 grey--text text--darken-2">
       {{ tenant.name }}
     </v-card-title>